note-sharing-service/requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.0.0
python-dotenv==1.0.0

